+++ streampipes-integration-tests/src/test/java/org/apache/streampipes/integration/adapters/MqttAdapterTLSTester.java
@@ -28,13 +28,9 @@ import org.apache.streampipes.integration.containers.MosquittoDevContainer;
 import org.apache.streampipes.integration.utils.Utils;
 import org.apache.streampipes.manager.template.AdapterTemplateHandler;
 import org.apache.streampipes.messaging.mqtt.MqttPublisher;
-import org.apache.streampipes.model.grounding.MqttTransportProtocol;
 import org.apache.streampipes.model.staticproperty.StaticPropertyAlternatives;
 import org.apache.streampipes.model.template.PipelineElementTemplate;
 
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.ArrayList;
@@ -109,32 +105,14 @@ public class MqttAdapterTLSTester extends AdapterTesterBase {
 
 
   @Override
-  public void publishEvents(List<Map<String, Object>> events) {
-    var publisher = getMqttPublisher();
-    var objectMapper = new ObjectMapper();
-
-    events.forEach(event -> {
-
-      try {
-        var serializedEvent = objectMapper.writeValueAsBytes(event);
-        publisher.publish(serializedEvent);
-      } catch (JsonProcessingException e) {
-        throw new RuntimeException(e);
-      }
-    });
-
-    publisher.disconnect();
+  public void publishEvents(List<Map<String, Object>> events) { 
+      var pub = getMqttPublisher();
+      MQTTPublisherUtils.publishEvents(pub, events);
   }
 
   @NotNull
   private MqttPublisher getMqttPublisher() {
-    MqttTransportProtocol mqttSettings = new MqttTransportProtocol(
-        mosquittoContainer.getBrokerHost(),
-        mosquittoContainer.getBrokerPort(),
-        TOPIC);
-    MqttPublisher publisher = new MqttPublisher(mqttSettings);
-    publisher.connect();
-    return publisher;
+    return MQTTPublisherUtils.getMqttPublisher(mosquittoContainer, TOPIC);
   }
 
   @Override
